requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==23.1
django_debug_toolbar==3.8.1
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-isort==6.0.0
flake8-plugin-utils==1.3.2
flake8-return==1.2.0
isort==5.12.0
mccabe==0.7.0
pep8-naming==0.13.3
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==1.0.0
sqlparse==0.4.3
tzdata==2023.3
